BVGraph: load OBL offsets even if they have a later modification time
Before this commit, BVGraph checked whether the OBL cache file was more
recent than the offsets file it was caching, to ensure that the cache
file was still valid.

However, file modification times are not a reliable way to check this.
When we copy the compressed graph from a remote server, the .obl file
will often be written first because it goes before .offsets in the
lexicographic order, and will then have an earlier ctime/mtime. We
therefore need to systematically touch(1) the .obl files after a copy.