BitStreamArcLabelledImmutableGraph: load offsets from OBL cache file
This commit adds the possibility of pre-computing the
EliasFanoMonotoneLongBigList which contains the labels of a
BitStreamArcLabelledImmutableGraph, and storing it as a .labelobl file.
At loading time, instead of having to deserialize the offsets from the
.labeloffsets file, it will simply load the .labelobl file directly,
which is a lot faster.

This class also needed a CLI helper to generate the .labelobl file from
the .labeloffsets file. However, previously the class did not expose a
main function at all. Because it would have been strange to have the
main function do only that, I implemented a main function which does a
similar thing to the BVGraph class:

- By default, it can store any ArcLabelledImmutableGraph into a
  BitStreamArcLabelledImmutableGraph, when called like this:

      $0 --underlying <underlying> <source> <destionation>

  where source is the basename of an ArcLabelledImmutableGraph,
  underlying is the underlying ImmutableGraph, and destination is the
  basename where the BitStreamArcLabelledImmutableGraph will be stored.

- Most importantly, it can also generate a .labelobl file from an
  existing BitStreamArcLabelledImmutableGraph, when called like this:

      $0 --list <source>

This mimicks the existing behavior of the BVGraph main, but adapted for
BitStreamArcLabelledImmutableGraph.